+++ addressbook-web-tests/src/test/java/com/lesson/addressbook/appmanager/ContactHelper.java
@@ -1,0 +1,64 @@
+package com.lesson.addressbook.appmanager;
+
+import com.lesson.addressbook.model.ContactData;
+import org.openqa.selenium.By;
+import org.openqa.selenium.firefox.FirefoxDriver;
+
+public class ContactHelper {
+    private FirefoxDriver wd;
+
+    public ContactHelper(FirefoxDriver wd) {
+        this.wd = wd;
+    }
+
+    public void submitCreationContact() {
+        wd.findElement(By.xpath("//div[@id='content']/form/input[21]")).click();
+    }
+
+    public void fillContactForm(ContactData contactData) {
+        wd.findElement(By.name("firstname")).click();
+        wd.findElement(By.name("firstname")).clear();
+        wd.findElement(By.name("firstname")).sendKeys(contactData.getFirstname());
+        wd.findElement(By.name("middlename")).click();
+        wd.findElement(By.name("middlename")).clear();
+        wd.findElement(By.name("middlename")).sendKeys(contactData.getMiddlename());
+        wd.findElement(By.name("lastname")).click();
+        wd.findElement(By.name("lastname")).clear();
+        wd.findElement(By.name("lastname")).sendKeys(contactData.getLastname());
+        wd.findElement(By.name("nickname")).click();
+        wd.findElement(By.name("nickname")).clear();
+        wd.findElement(By.name("nickname")).sendKeys(contactData.getNickname());
+        wd.findElement(By.name("company")).click();
+        wd.findElement(By.name("company")).clear();
+        wd.findElement(By.name("company")).sendKeys(contactData.getCompany());
+        wd.findElement(By.name("address")).click();
+        wd.findElement(By.name("address")).clear();
+        wd.findElement(By.name("address")).sendKeys(contactData.getAddress());
+        wd.findElement(By.name("home")).click();
+        wd.findElement(By.name("home")).clear();
+        wd.findElement(By.name("home")).sendKeys(contactData.getHomephone());
+        wd.findElement(By.name("email")).click();
+        wd.findElement(By.name("email")).clear();
+        wd.findElement(By.name("email")).sendKeys(contactData.getEmail());
+        wd.findElement(By.name("homepage")).click();
+        wd.findElement(By.name("homepage")).clear();
+        wd.findElement(By.name("homepage")).sendKeys(contactData.getHomepage());
+    }
+
+    public void goToAddNewContact() {
+        wd.findElement(By.linkText("add new")).click();
+    }
+
+    public void closeAlertWindow() {
+        wd.switchTo().alert().accept();
+    }
+
+    public void deleteContact() {
+
+        wd.findElement(By.xpath(".//*[@value='Delete']")).click();
+    }
+
+    public void selectContact() {
+        wd.findElement(By.xpath(".//*[@class=\"center\"][1]")).click();
+    }
+}
